object_detection_app_v1/requirements.txt
streamlit>=1.32
ultralytics>=8.0.0
opencv-python-headless>=4.8.0.76
Pillow
numpy
